% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{adjMat2Moral}
\alias{adjMat2Moral}
\title{Convert Adjacency matrix to Moral (Undirected Conditional Dependence) Network}
\usage{
adjMat2Moral(adj.mat)
}
\arguments{
\item{adj.mat}{An adjacency matrix}
}
\value{
An object of class bn where all edges are undirected
}
\description{
Convert Adjacency matrix to Moral (Undirected Conditional Dependence) Network
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

