% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{convert2BinaryMat}
\alias{convert2BinaryMat}
\title{Convert to Binary Matrix}
\usage{
convert2BinaryMat(mat)
}
\description{
Converts a matrix to a binary matrix, where 0 values remain 0 values, and non-zero values
become 1.
}

