% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{getPrecisionBasis}
\alias{getPrecisionBasis}
\title{Get the precision matrix structure for a Gaussian joint distribution based on a Bayesian
network structure}
\usage{
getPrecisionBasis(net)
}
\arguments{
\item{net}{A object of class \code{bn}.}
}
\value{
A symetric matrix.
}
\description{
Converts a Bayesian network to a binary matrix where 1's correspond to non-zero elements
in the precision matrix.
}

