% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmsim.R
\docType{data}
\name{melancon_boot}
\alias{melancon_boot}
\title{Bootstrap Learning Performed on a Gaussian Bayesian Network.}
\format{A list with 2 items}
\usage{
data(melancon_boot)
}
\description{
The 15 node network structure was simulated with Melancon's and Philippe's
 Uniform Random Acyclic Digraphs algorithm.  This produces a highly connected DAG.
 See \code{?random.graph} for more information.
}
\details{
\itemize{
  \item truth. An object of \code{bn.fit}, a simulated structure and parameter set.
  \item boot.  An object of \code{bn.strength} and \code{data.frame}, results of
  model averaging applied to data simulated from truth.
}
}
\keyword{datasets}

