% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmsim.R
\docType{data}
\name{ordered_boot}
\alias{ordered_boot}
\title{Bootstrap Learning Performed on a Gaussian Bayesian Network.}
\format{A list with 2 items}
\usage{
data(ordered_boot)
}
\description{
The 40 node network structure was simulated with full ordering based generation.
 This produces a sparsely connected DAG.  See \code{?random.graph} for more information.
}
\details{
\itemize{
  \item truth. An object of \code{bn.fit}, a simulated structure and parameter set.
  \item boot.  An object of \code{bn.strength} and \code{data.frame}, results of
  model averaging applied to data simulated from truth.
}
}
\keyword{datasets}

