% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{randomNet}
\alias{randomNet}
\title{Random Graph Simulation}
\usage{
randomNet(num.vars, method)
}
\arguments{
\item{num.vars}{Number of variables}

\item{method}{A character string, possible values are \code{ordered},
\code{ic-dag}, and \code{melancon}. See \code{?random.graph} in \pkg{bnlearn}}
}
\value{
an object of class \code{bn}.
}
\description{
This is a wrapper for the bnlearn packages *random.graph* which gives a
single graph with node names as numbers prefixed with "V", to match default
naming of variables in data.frames.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

