% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simBNFromCovAndDAG}
\alias{simBNFromCovAndDAG}
\title{Simulate a Bayesian network from a DAG and covariance matrix}
\usage{
simBNFromCovAndDAG(dag, cov.mat)
}
\arguments{
\item{dag}{A object of class bn, with no undirected edges.}

\item{cov.mat}{A positive definate matrix.}
}
\value{
An object of class \code{bn.fit}
}
\description{
Given a covariance matrix and a directed acyclic graph structure,
simulates a fully parameterized Guassian Bayesian network consistent
with the covariance matrix.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

