% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simBNFromCovMat}
\alias{simBNFromCovMat}
\title{Simulate a Bayesian network from a covariance matrix}
\usage{
simBNFromCovMat(cov.mat)
}
\arguments{
\item{cov.mat}{A positive definate matrix.}
}
\value{
An object of class \code{bn.fit}. If no consistend extention is found,
 NULL is returned.
}
\description{
Given a covariance matrix, simulates a fully parameterized Guassian Bayesian network
consistent with the covariance matrix.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

