% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simDAGFromCovMat}
\alias{simDAGFromCovMat}
\title{Simulate a Gaussian Bayesian network structure based on a covariance matrix.}
\usage{
simDAGFromCovMat(cov.mat)
}
\arguments{
\item{cov.mat}{A covariance matrix (positive definate matrix).}
}
\value{
An objective of class bn.  If no consistent extention is found, NULL is returned.
}
\description{
This function attempts to find a directed acyclic structure consistent with the
conditional independencies given by the covariance matrix under Gaussian assumptions.
If such a consistent extention cannot be found, then NULL is returned.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
  \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

