% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simGaussianNet}
\alias{simGaussianNet}
\title{Simulate a fully specified Gaussian Bayesian network}
\usage{
simGaussianNet(num.vars, rng = c(0.5, 1.5), method = "melancon")
}
\arguments{
\item{num.vars}{Number of variables/nodes}

\item{rng}{The desired range of the eigen values in the covariance matrix.}

\item{method}{A character string, possible values are \code{ordered},
\code{ic-dag}, and \code{melancon}. See \code{?random.graph} in \pkg{bnlearn}}
}
\value{
An object of class \code{bn.fit}
}
\description{
Simulates a Gaussian Bayesian network by simulating a network structure and
a random covariance matrix consistent with the I-map of the network structure,
and then calculating network parameters based on the covariance matrix.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

