% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simMVNData}
\alias{simMVNData}
\title{Simulation of Multivariate Normal dataset}
\usage{
simMVNData(num.vars, num.obs)
}
\arguments{
\item{num.vars}{The number of variables.}

\item{num.obs}{The number of observations.}
}
\value{
A data frame of multivariate normal observations.
}
\description{
Creates a Multivariate Normal Dataset via random positive definate matrix generation.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

