% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simNetCovariance}
\alias{simNetCovariance}
\title{Simulate a covariance matrix based on a Bayesian network structure}
\usage{
simNetCovariance(net, rng)
}
\arguments{
\item{net}{A Bayesian network structure, an object of class bn.}

\item{rng}{The desired range of the eigen values in the covariance matrix.}
}
\value{
A simulated the covariance matrix consistent with the net argument. Simulated
under the Gaussian assumption.
}
\description{
Given a Bayesian network structure, simulates a covariance matrix consistent
with the network structure, based on Gaussian assumptions.
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

