% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simSparsePrecision}
\alias{simSparsePrecision}
\title{Simulate a Sparce Precision Matrix}
\usage{
simSparsePrecision(basis, rng)
}
\arguments{
\item{basis}{A binary matrix where 1's correspond to non-zero elements
in the precision matrix. This is the output of the \code{getPrecisionBasis} function.}

\item{rng}{The desired range of the eigen values for the matrix.}
}
\value{
A positive definate matrix of the same dimension as the basis argument.
}
\description{
Simulate a Sparce Precision Matrix
}
\seealso{
\code{\link{simGaussianNet}}, \code{\link{simBNFromCovMat}},
 \code{\link{simDAGFromCovMat}}, \code{\link{simNetCovariance}}, \code{\link{simMVNData}},
 \code{\link{randomNet}}, \code{\link{simSparsePrecision}}, \code{\link{simDAGFromCovMat}}
}

